+++ package.json
@@ -29,6 +29,7 @@
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@astrojs/starlight-tailwind": "^2.0.3",
+    "@astrojs/tailwind": "^5.1.2",
     "astro": "^4.15.3",
     "classnames": "^2.5.1",
     "react": "^18.3.1",
