+++ tailwind.config.js
@@ -3,7 +3,11 @@ import colors from "tailwindcss/colors";
 
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ["./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}"],
+  content: [
+    "./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}",
+    "./public/**/*.html",
+  ],
+  plugins: [starlightPlugin()],
   theme: {
     extend: {
       colors: {
@@ -20,5 +24,37 @@ export default {
       },
     },
   },
-  plugins: [starlightPlugin()],
+};
+
+module.exports = {
+  theme: {
+    screens: {
+      sm: "480px",
+      md: "768px",
+      lg: "976px",
+      xl: "1440px",
+    },
+    colors: {
+      blue: "#1fb6ff",
+      pink: "#ff49db",
+      orange: "#ff7849",
+      green: "#13ce66",
+      "gray-dark": "#273444",
+      gray: "#8492a6",
+      "gray-light": "#d3dce6",
+    },
+    fontFamily: {
+      sans: ["Graphik", "sans-serif"],
+      serif: ["Merriweather", "serif"],
+    },
+    extend: {
+      spacing: {
+        128: "32rem",
+        144: "36rem",
+      },
+      borderRadius: {
+        "4xl": "2rem",
+      },
+    },
+  },
 };
